refactor and first spring form

+++ BankingSystem/src/main/java/com/tsmiecho/banking/pojo/Itransfer.java
@@ -1,10 +1,12 @@
 package com.tsmiecho.banking.pojo;
 
+import java.math.BigDecimal;
+
 public interface Itransfer {
 	
 	Account getSenderAccount();
 	
-	double getBalanceAfterTransactionSender();
+	BigDecimal getBalanceAfterTransactionSender();
 	
-	void setBalanceAfterTransactionSender(double balanceAfterTransactionSender);
+	void setBalanceAfterTransactionSender(BigDecimal balanceAfterTransactionSender);
 }
